requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2.2
djangorestframework==3.16.0
djangorestframework-simplejwt
django-cors-headers
iniconfig==2.1.0
kombu==5.5.4
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
Pygments==2.19.1
pytest==8.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.2.0
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
drf-yasg
